feat(stats): make stats work again
In order to cleanup, you can run the following, though it's not required:

```
rm -rf /opt/isard-local/grafana
rm -rf /opt/isard-local/stats
```

This commit also removes the `grafana.db` binary, and now it's
constructed in the first run. Also, it makes Grafana read only. This
allows running the container as the grafana user, and enables automatic
dashboard updates. Keep in mind that you won't be able to save changes
in Grafana!