fix(authentication)!: fix the LDAP group authentication
BREAKING CHANGE: Now, if the LDAP auto register is enabled, the groups
will be created as external. This means that if there's already a LDAP
group that you created manually with the same name in IsardVDI and LDAP,
is not going to have the new users created there, it will create those
in a new group that will be created automatically. With the following
script, you can migrate the DB to reuse the current groups with the new
system. You need to execute it inside `isard-engine` (`docker exec -it
isard-engine /bin/bash`)

```python
from engine.services.db import *

r_conn = new_rethink_connection()
rtable = r.table("groups")

l_groups = list(rtable.pluck("id", "name").run(r_conn))

 ########
 # vvvv #
l_groups_ldap = [  ] # HERE THERE SHOULD BE ALL THE LDAP GROUPS (THE VALUE THAT'S IN NAME)
 # ^^^^ #
 ########

ids_update = {a["id"]: a["name"] for a in l_groups if a["name"] in l_groups_ldap}
for k, v in ids_update.items():
    rtable.get(k).update(
        {
            "external_app_id": "provider-ldap",
            "external_gid": v,
            "description": f"ldap {v}",
        },
    ).run(r_conn)

print("Migration success! :)")
```