fix: add missing attribute to enable auto-registration
The auto-registration feature was malfunctioning due to the absence of `l.cfg.FieldGroup`
in the `attributes` array. This omission resulted in `ExternalGID` always being empty.
To resolve this, code was added to include `l.cfg.FieldGroup` in `attributes` when
`l.AutoRegister()` is true, specifically before creating `req := ldap.NewSearchRequest`.
This ensures the necessary attribute is present for successful auto-registration.

Refs: #609