feat(orchestrator): don't remove only forced when hypervisor is created
This feature solves an issue we're encountering with the orchestrator.
Sometimes, orchestrator creates a new hypervisor but when the hypervisor
comes online is no longer needed. With the previous behaviour, it took
one iteration for the orchestrator to decide it was no longer needed,
but between the iteration that creates it and the one that decides is no
longer needed, a desktop could be started in this new hypervisor and
then it no longer could be destroyed.

With the behaviour that this change introduces, it will take a full
iteration for the orchestrator to remove it from only forced, so:
a. If the hypervisor is still needed, it will remove it from only forced
b. If the hypervisor is no longer needed, it will be removed, since it
won't have any desktop started