feat(authentication): added guess category for SAML authentication
Added a guess category parameter in config that defines wether the
categories will be automatically retrieved from the SAML user. If guess
category is enabled and one category is detected the user will be logged
into that category, if more than one category is detected a token will
be returned with the users categories.