feat(api): user migration import endpoint
Added a new API endpoint that checks that the given migration token
belongs to a existing migration and calls authentication to check
whether the token is valid or not. If the token is not valid the
migration will be removed from the db, otherwise the target user is
set and the migration status is updated.