fix(api,scheduler): nonpersistent desktops stop schedule delete job
When stopping a nonpersistent desktop an scheduler job is created in
order to delete the desktop in the configured minutes. If the desktop is
started the job is removed until it's stopped. Then, another job is
scheduled to delete it. This process will repeat until the nonpersistent desktop is
deleted.